Remove `<SelectIDE />`
